+++ R/type-date-time.R
@@ -61,9 +61,10 @@ new_duration <- function(x = double(), units = c("secs", "mins", "hours", "days"
 }
 
 #' @export
-vec_is_vector.POSIXlt <- function(x) {
-  TRUE
+vec_proxy.POSIXlt <- function(x) {
+  new_data_frame(unclass(x))
 }
+
 
 # Print ------------------------------------------------------------------