+++ src/init.c
@@ -31,7 +31,7 @@ extern SEXP vctrs_match(SEXP, SEXP);
 extern SEXP vctrs_duplicated_any(SEXP);
 extern SEXP vctrs_size(SEXP);
 extern SEXP vctrs_is_unspecified(SEXP);
-extern SEXP vctrs_typeof(SEXP);
+extern SEXP vctrs_typeof(SEXP, SEXP);
 extern SEXP vctrs_is_vector(SEXP);
 extern SEXP vctrs_type2(SEXP);
 extern SEXP vctrs_typeof2(SEXP, SEXP);
@@ -68,7 +68,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_equal_na",                   (DL_FUNC) &vctrs_equal_na, 1},
   {"vctrs_compare",                    (DL_FUNC) &vctrs_compare, 3},
   {"vctrs_match",                      (DL_FUNC) &vctrs_match, 2},
-  {"vctrs_typeof",                     (DL_FUNC) &vctrs_typeof, 1},
+  {"vctrs_typeof",                     (DL_FUNC) &vctrs_typeof, 2},
   {"vctrs_init",                       (DL_FUNC) &vctrs_init, 1},
   {"vctrs_is_vector",                  (DL_FUNC) &vctrs_is_vector, 1},
   {"vctrs_type2",                      (DL_FUNC) &vctrs_type2, 2},
@@ -89,6 +89,7 @@ void R_init_vctrs(DllInfo *dll)
 
 
 void vctrs_init_cast(SEXP ns);
+void vctrs_init_data(SEXP ns);
 void vctrs_init_size(SEXP ns);
 void vctrs_init_type2(SEXP ns);
 void vctrs_init_types(SEXP ns);
@@ -96,6 +97,7 @@ void vctrs_init_utils(SEXP ns);
 
 SEXP vctrs_init(SEXP ns) {
   vctrs_init_cast(ns);
+  vctrs_init_data(ns);
   vctrs_init_size(ns);
   vctrs_init_type2(ns);
   vctrs_init_types(ns);