+++ tests/testthat/helper-s3.R
@@ -9,3 +9,10 @@ unrownames <- function(x) {
 scoped_global_bindings <- function(..., .frame = caller_env()) {
   scoped_bindings(..., .env = global_env(), .frame = .frame)
 }
+
+new_proxy <- function(x) {
+  structure(list(env(x = x)), class = "vctrs_proxy")
+}
+proxy_deref <- function(x) {
+  x[[1]]$x
+}